start algorithms - combinations, arrangements
