+++ .gitignore
@@ -1,3 +1,7 @@
+# working with intelliJ
+.idea
+JAVAresources.iml
+
 # Compiled class file
 *.class
